requirements.txt
backcall==0.2.0
certifi==2020.12.5
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
colorama==0.4.1
cycler==0.11.0
dash==1.20.0
dash-bootstrap-components==0.12.1a4
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
decorator==5.1.1
Flask==2.1.2
Flask-AppBuilder==4.1.2
Flask-Babel==2.0.0
Flask-Caching==2.0.0
Flask-Compress==1.12
Flask-JWT-Extended==4.4.2
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
flower==1.1.0
folium==0.11.0
Django==2.2.5
Faker==2.0.1
future==0.17.1
image==1.5.27
ipython==7.8.0
ipython-genutils==0.2.0
itsdangerous==2.0
JayDeBeApi==1.1.1
jedi==0.15.1
Jinja2==3.1.2
joblib
jpype1
kiwisolver
MarkupSafe
matplotlib==3.1.1
numpy
pandas
parso==0.5.1
pathlib==1.0.1
pickleshare==0.7.5
Pillow==6.1.0
pipenv==2018.11.26
plotly==5.10.0
prompt-toolkit==2.0.9
psycopg2==2.8.3
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
retrying==1.3.3
Rtree==0.8.3
scipy
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==2.0
gunicorn==20.0.4
folium
cufflinks
matplotlib_venn
unidecode
wordcloud
flask_mail
mysqlclient
mysql-connector-python==8.0.29
pymysql
cryptography
